specify actuator healthcheck endpoint refactor package name
