+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+management.endpoints.web.base-path=/disqualified-officers-search-consumer/
+management.endpoints.web.path-mapping.health=healthcheck
